Add helm dependency management
Minor code refactoring
