+++ pkg/plugins/resources/helm/source.go
@@ -1,10 +1,6 @@
 package helm
 
 import (
-	"fmt"
-	"io/ioutil"
-	"net/http"
-
 	"github.com/sirupsen/logrus"
 	"github.com/updatecli/updatecli/pkg/core/result"
 )
@@ -12,29 +8,7 @@ import (
 // Source return the latest version
 func (c *Chart) Source(workingDir string) (string, error) {
 
-	URL := fmt.Sprintf("%s/index.yaml", c.spec.URL)
-
-	req, err := http.NewRequest("GET", URL, nil)
-
-	if err != nil {
-		return "", err
-	}
-
-	res, err := http.DefaultClient.Do(req)
-
-	if err != nil {
-		return "", err
-	}
-
-	defer res.Body.Close()
-
-	body, err := ioutil.ReadAll(res.Body)
-
-	if err != nil {
-		return "", err
-	}
-
-	index, err := loadIndex(body)
+	index, err := c.GetRepoIndexFile()
 
 	if err != nil {
 		return "", err
