+++ pkg/plugins/resources/helm/target.go
@@ -1,24 +1,19 @@
 package helm
 
 import (
-	"fmt"
-	"io/ioutil"
-	"os"
+	"bytes"
 	"path/filepath"
-	"strings"
 
-	"github.com/Masterminds/semver/v3"
 	"github.com/sirupsen/logrus"
 	"github.com/updatecli/updatecli/pkg/core/pipeline/scm"
 	"github.com/updatecli/updatecli/pkg/plugins/resources/yaml"
-	helm "helm.sh/helm/v3/pkg/chart"
-
-	YAML "sigs.k8s.io/yaml"
 )
 
 // Target updates helm chart, it receives the default source value and a dryrun flag
 // then return if it changed something or failed
 func (c *Chart) Target(source string, dryRun bool) (changed bool, err error) {
+	var out bytes.Buffer
+
 	err = c.ValidateTarget()
 	if err != nil {
 		return false, err
@@ -48,16 +43,21 @@ func (c *Chart) Target(source string, dryRun bool) (changed bool, err error) {
 		return false, nil
 	}
 
-	// Reset requirements.lock if we modified the file 'requirements.yaml'
-	if strings.Compare(c.spec.File, "requirements.yaml") == 0 && !dryRun {
-		_, err := c.UpdateRequirements(filepath.Join(c.spec.Name, "requirements.lock"))
-		if err != nil {
-			return false, err
-		}
+	// Update Chart.yaml file new Chart Version and appVersion if needed
+	err = c.MetadataUpdate(c.spec.Name, dryRun)
+	if err != nil {
+		return false, err
 	}
 
-	// Update Chart.yaml file new Chart Version and appVersion if needed
-	err = c.UpdateMetadata(filepath.Join(c.spec.Name, "Chart.yaml"), dryRun)
+	err = c.RequirementsUpdate(c.spec.Name)
+	if err != nil {
+		return false, err
+	}
+
+	err = c.DependencyUpdate(&out, "")
+
+	logrus.Infof("%s", out.String())
+
 	if err != nil {
 		return false, err
 	}
@@ -69,6 +69,8 @@ func (c *Chart) Target(source string, dryRun bool) (changed bool, err error) {
 // then return if it changed something or failed
 func (c *Chart) TargetFromSCM(source string, scm scm.ScmHandler, dryRun bool) (
 	changed bool, files []string, message string, err error) {
+
+	var out bytes.Buffer
 	err = c.ValidateTarget()
 	if err != nil {
 		return false, files, message, err
@@ -79,7 +81,6 @@ func (c *Chart) TargetFromSCM(source string, scm scm.ScmHandler, dryRun bool) (
 		Key:  c.spec.Key,
 	}
 	if len(c.spec.Value) == 0 {
-		c.spec.Value = source
 		c.spec.Value = source
 	} else {
 		yamlSpec.Value = c.spec.Value
@@ -97,158 +98,27 @@ func (c *Chart) TargetFromSCM(source string, scm scm.ScmHandler, dryRun bool) (
 		return false, files, message, nil
 	}
 
-	if strings.Compare(c.spec.File, "requirements.yaml") == 0 {
-		found, err := c.UpdateRequirements(filepath.Join(scm.GetDirectory(), c.spec.Name, "requirements.lock"))
-		if err != nil {
-			return false, files, message, err
-		}
-		if found {
-			files = append(files, filepath.Join(c.spec.Name, "requirements.lock"))
+	chartPath := filepath.Join(scm.GetDirectory(), c.spec.Name)
 
-		}
-	}
-
-	err = c.UpdateMetadata(filepath.Join(scm.GetDirectory(), c.spec.Name, "Chart.yaml"), dryRun)
+	err = c.MetadataUpdate(chartPath, dryRun)
 	if err != nil {
 		return false, files, message, err
 	}
-	files = append(files, filepath.Join(c.spec.Name, "Chart.yaml"))
+
+	err = c.RequirementsUpdate(chartPath)
+	if err != nil {
+		return false, files, message, err
+	}
+
+	err = c.DependencyUpdate(&out, chartPath)
+
+	if err != nil {
+		return false, files, message, err
+	}
+
+	logrus.Debug(out.String())
+
+	files = append(files, chartPath)
 
 	return changed, files, message, err
-}
-
-// UpdateRequirements test if we are updating the file requirements.yaml
-// if it's the case then we also have to delete and recreate the file
-// requirements.lock
-func (c *Chart) UpdateRequirements(lockFilename string) (bool, error) {
-	if strings.Compare(c.spec.File, "requirements.yaml") != 0 {
-		return false, fmt.Errorf("No need to cleanup requirements.lock")
-	}
-
-	f, err := os.Stat(lockFilename)
-
-	if os.IsExist(err) && !f.IsDir() {
-		err = os.Remove(lockFilename)
-		if err != nil {
-			return false, err
-		}
-		return true, nil
-	} else if os.IsNotExist(err) {
-		return false, nil
-	}
-
-	logrus.Debugf("Something went wrong with lock file %v", lockFilename)
-	return false, fmt.Errorf("Something unexpected happened")
-
-}
-
-// UpdateMetadata updates a metadata if necessary and it bump the ChartVersion
-func (c *Chart) UpdateMetadata(metadataFilename string, dryRun bool) error {
-	md := helm.Metadata{}
-
-	file, err := os.Open(metadataFilename)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	data, err := ioutil.ReadAll(file)
-
-	if err != nil {
-		return err
-	}
-
-	err = YAML.Unmarshal(data, &md)
-
-	if err != nil {
-		return err
-	}
-
-	if len(md.AppVersion) > 0 && c.spec.AppVersion {
-		logrus.Infof("AppVersion updated from %s to %s\n", md.AppVersion, c.spec.Value)
-		md.AppVersion = c.spec.Value
-	}
-
-	// Init Chart Version if not set yet
-	if len(md.Version) == 0 {
-		md.Version = "0.0.0"
-	}
-
-	oldVersion := md.Version
-
-	for _, inc := range strings.Split(c.spec.VersionIncrement, ",") {
-		v, err := semver.NewVersion(md.Version)
-		if err != nil {
-			return err
-		}
-
-		switch inc {
-		case MAJORVERSION:
-			md.Version = v.IncMajor().String()
-		case MINORVERSION:
-			md.Version = v.IncMinor().String()
-		case PATCHVERSION:
-			md.Version = v.IncPatch().String()
-		default:
-			logrus.Errorf("Wrong increment rule %q.", inc)
-		}
-	}
-
-	logrus.Infof("Chart Version updated from %q to %q\n", oldVersion, md.Version)
-
-	if err != nil {
-		return err
-	}
-
-	if !dryRun {
-		data, err := YAML.Marshal(md)
-		if err != nil {
-			return err
-		}
-
-		file, err := os.Create(metadataFilename)
-		if err != nil {
-			return err
-		}
-
-		defer file.Close()
-
-		_, err = file.Write(data)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-//ValidateTarget ensure that target required parameter are set
-func (c *Chart) ValidateTarget() error {
-	if len(c.spec.File) == 0 {
-		c.spec.File = "values.yaml"
-	}
-
-	if len(c.spec.Name) == 0 {
-		return fmt.Errorf("Parameter name required")
-	}
-
-	if len(c.spec.Key) == 0 {
-		return fmt.Errorf("Parameter key required")
-	}
-
-	if len(c.spec.VersionIncrement) == 0 {
-		c.spec.VersionIncrement = MINORVERSION
-	}
-
-	for _, inc := range strings.Split(c.spec.VersionIncrement, ",") {
-
-		if inc != MAJORVERSION &&
-			inc != MINORVERSION &&
-			inc != PATCHVERSION &&
-			inc != "" {
-			return fmt.Errorf("Unrecognized increment rule %q. accepted values are a comma separated list of [major,minor,patch]", inc)
-		}
-	}
-
-	return nil
 }
